third_party/blink/web_tests/fast/forms/search/intrinsic-search-width-with-decoration-border-padding-expected.txt

PASS document.getElementById("search1").offsetWidth is document.getElementById("search2").offsetWidth
PASS successfullyParsed is true

TEST COMPLETE

 
